Updated modify_reach_curator_xlsx.ipynb & analyze_curated_reach_results.ipynb
Added 3 changes to both files (same changes for both files)

(1) Changed the hard coding of the mouse = 'reach10' in the class instance. and made it strip the mouse name from the root_file path

(2) Added a Kernal KILL switch. if you try to run the notebook over a session that already has analyzed data frames. it will exit() / kill the kernal.
> you have to skip this cell to overwrite the data

(3) Made the path creation in the get_mouse_session_files(self): function for the (txt_file, and xlsx_orig) use the main_path instead of hard coding the first part of that path in